python/scripts/preprocessing/example_requirements.txt
grpcio
protobuf
pyarrow
